[MOB-2931] Remove V1 quotes and availability for V2 (#51)
* Tests fixed

* Tests fixed

* Version bump

* Uploaded canary version

* Remove V2

* Remove more V2's

* Fix error in TestData
