+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,7 +1,6 @@
-#Tue Apr 25 17:27:27 BST 2017
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.4.1-bin.zip
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.4.1-all.zip
 kotlin.coroutines=enable
