+++ sdk/build.gradle
@@ -72,7 +72,7 @@ android {
 publish {
     def groupProjectID = 'com.karhoo'
     def artifactProjectID = 'nsdk-canary'
-    def publishVersionID = '1.0.35.28092020-1'
+    def publishVersionID = '1.0.35.30092020-2'
     def repoNameID = 'Android'
 
     userOrg = 'flit-tech'
