+++ sdk/src/main/java/com/karhoo/sdk/api/model/Vehicles.kt
@@ -1,12 +1,11 @@
 package com.karhoo.sdk.api.model
 
 import android.os.Parcelable
-
 import com.google.gson.annotations.SerializedName
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class Vehicles(@SerializedName("id") val id: String? = null,
+data class Vehicles(@SerializedName("id") val id: String = "",
                     @SerializedName("status") val status: String? = null,
-                    @SerializedName("category_names") val categoryNames: List<String> = emptyList(),
-                    @SerializedName("quote_items") val vehicles: List<Quote> = emptyList()) : Parcelable
+                    @SerializedName("availability") val availability: Availability = Availability(),
+                    @SerializedName("quotes") val quotes: List<Quote> = emptyList()) : Parcelable
