+++ sdk/src/main/java/com/karhoo/sdk/api/network/header/KarhooHeaders.kt
@@ -17,7 +17,7 @@ class KarhooHeaders(private val credentialsManager: CredentialsManager) : Header
     override fun generateCorrelationId(method: String): String {
         return when (method) {
             APITemplate.BOOKING_WITH_NONCE_METHOD -> configuredQuoteId()
-            APITemplate.QUOTE_REQUEST_METHOD -> updatedQuoteId()
+            APITemplate.QUOTES_REQUEST_METHOD -> updatedQuoteId()
             else -> ANDROID + UUID.randomUUID()
         }
     }
