+++ sdk/src/main/java/com/karhoo/sdk/api/network/request/QuotesRequest.kt
@@ -5,6 +5,7 @@ import com.google.gson.annotations.SerializedName
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class QuotesRequest(@SerializedName("origin_place_id") val originPlaceId: String?,
-                         @SerializedName("destination_place_id") val destinationPlaceId: String?,
-                         @SerializedName("local_time_of_pickup") val dateScheduled: String?) : Parcelable
+data class QuotesRequest(@SerializedName("origin") val origin: QuoteRequestPoint,
+                         @SerializedName("destination") val destination: QuoteRequestPoint,
+                         @SerializedName("local_time_of_pickup") val dateScheduled: String?) :
+        Parcelable
