+++ sdk/src/main/java/com/karhoo/sdk/di/ApplicationComponent.kt
@@ -3,7 +3,6 @@ package com.karhoo.sdk.di
 import com.karhoo.sdk.api.datastore.user.KarhooUserStore
 import com.karhoo.sdk.api.service.address.KarhooAddressService
 import com.karhoo.sdk.api.service.auth.KarhooAuthService
-import com.karhoo.sdk.api.service.availability.KarhooAvailabilityService
 import com.karhoo.sdk.api.service.config.KarhooConfigService
 import com.karhoo.sdk.api.service.drivertracking.KarhooDriverTrackingService
 import com.karhoo.sdk.api.service.fare.KarhooFareService
@@ -16,7 +15,6 @@ import dagger.Component
 @Component(modules = [
     NetworkModule::class,
     UserModule::class,
-    InteractorModule::class,
     AnalyticsModule::class])
 interface ApplicationComponent {
 
@@ -33,8 +31,6 @@ interface ApplicationComponent {
     fun inject(driverTrackingService: KarhooDriverTrackingService)
 
     fun inject(quotesService: KarhooQuotesService)
-
-    fun inject(availabilityService: KarhooAvailabilityService)
 
     fun inject(configService: KarhooConfigService)
